feat: have 2 run actions - run latest and run (#172)
